moved WebResourceContext to web and api cleanup
