+++ crabzilla-pg-client/src/main/java/io/github/crabzilla/pgc/projections.kt
@@ -2,9 +2,9 @@ package io.github.crabzilla.pgc
 
 import io.github.crabzilla.core.DomainEvent
 import io.github.crabzilla.core.EVENT_SERIALIZER
+import io.github.crabzilla.core.PgcEventBusChannels
 import io.github.crabzilla.core.UnitOfWork
 import io.github.crabzilla.core.UnitOfWorkEvents
-import io.github.crabzilla.internal.PgcEventBusChannels
 import io.vertx.core.Future
 import io.vertx.core.Future.failedFuture
 import io.vertx.core.Promise
